LPS-83992 Move out of the kernel.util dumpster
