+++ portal-impl/src/com/liferay/portlet/internal/PortletAsyncContextImpl.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portlet.internal;
 
-import com.liferay.portal.kernel.util.PortletAsyncScopeManager;
-import com.liferay.portal.kernel.util.PortletAsyncScopeManagerFactory;
+import com.liferay.portal.kernel.portlet.async.PortletAsyncScopeManager;
+import com.liferay.portal.kernel.portlet.async.PortletAsyncScopeManagerFactory;
 import com.liferay.portal.kernel.util.ServiceProxyFactory;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.WebKeys;
