+++ portal-impl/src/com/liferay/portlet/internal/PortletAsyncScopingRunnable.java
@@ -16,8 +16,8 @@ package com.liferay.portlet.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.util.PortletAsyncScopeManager;
-import com.liferay.portal.kernel.util.PortletAsyncScopeManagerFactory;
+import com.liferay.portal.kernel.portlet.async.PortletAsyncScopeManager;
+import com.liferay.portal.kernel.portlet.async.PortletAsyncScopeManagerFactory;
 import com.liferay.portlet.PortletAsyncListenerAdapter;
 
 import java.io.IOException;
