+++ portal-kernel/src/com/liferay/portal/kernel/portlet/async/PortletAsyncScopeManagerFactory.java
@@ -12,18 +12,23 @@
  * details.
  */
 
-package com.liferay.portal.kernel.util;
+package com.liferay.portal.kernel.portlet.async;
 
 import aQute.bnd.annotation.ProviderType;
+
+import javax.portlet.PortletConfig;
+import javax.portlet.ResourceRequest;
+import javax.portlet.ResourceResponse;
 
 /**
  * @author Neil Griffin
  */
+@FunctionalInterface
 @ProviderType
-public interface PortletAsyncScopeManager {
+public interface PortletAsyncScopeManagerFactory {
 
-	public void activateScopeContexts();
-
-	public void deactivateScopeContexts();
+	public PortletAsyncScopeManager getPortletAsyncScopeManager(
+		ResourceRequest resourceRequest, ResourceResponse resourceResponse,
+		PortletConfig portletConfig);
 
 }
